Move Parse and TryParse to CodeOwnersFile
Parsing a CODEOWNERS file now starts from the type it produces, the way
XDocument.Parse and SemanticVersion.Parse do, instead of from a separate
static class that exists only to hold three methods:

    CodeOwnersFile file = CodeOwnersFile.Parse(content);

CodeOwnersParser keeps the same three methods, forwarding to
CodeOwnersFile and marked [Obsolete], so callers migrating from 5.x get a
warning pointing at the replacement rather than a missing type.

The parsing engine moves out of CodeOwnersParser into its own internal
CodeOwnersParserContext type, so the public entry points and the obsolete
facade share one implementation.
